refactor fetchPortalConfiguration into EnterpriseApp
