+++ package.json
@@ -85,8 +85,8 @@
     "redux-mock-store": "^1.5.1",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.20.2",
-    "webpack": "^4.12.0",
-    "webpack-cli": "^2.0.10",
+    "webpack": "^4.25.1",
+    "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.0",
     "webpack-merge": "^4.1.1"
   },
