+++ src/containers/CodeManagementPage/index.jsx
@@ -3,7 +3,6 @@ import { connect } from 'react-redux';
 import CodeManagement from '../../components/CodeManagement';
 
 import { fetchCouponOrders, clearCouponOrders } from '../../data/actions/coupons';
-import { fetchPortalConfiguration } from '../../data/actions/portalConfiguration';
 
 const mapStateToProps = state => ({
   loading: state.coupons.loading,
@@ -13,9 +12,6 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-  fetchPortalConfiguration: (slug) => {
-    dispatch(fetchPortalConfiguration(slug));
-  },
   fetchCouponOrders: () => {
     dispatch(fetchCouponOrders());
   },
