+++ src/containers/EnterpriseApp/index.jsx
@@ -2,6 +2,8 @@ import { connect } from 'react-redux';
 
 import EnterpriseApp from '../../components/EnterpriseApp';
 
+import { fetchPortalConfiguration } from '../../data/actions/portalConfiguration';
+
 const mapStateToProps = (state) => {
   const enterpriseListState = state.table['enterprise-list'] || {};
 
@@ -11,4 +13,10 @@ const mapStateToProps = (state) => {
   };
 };
 
-export default connect(mapStateToProps)(EnterpriseApp);
+const mapDispatchToProps = dispatch => ({
+  fetchPortalConfiguration: (slug) => {
+    dispatch(fetchPortalConfiguration(slug));
+  },
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(EnterpriseApp);
